requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
openpyxl>=3.1.0
PyYAML>=6.0.0
